"drag state" classes aren't added in touch interfaces
This is because there is a method that prevents course pointer
interfaces (i.e., touch) from text selecting the UI. But the classes
are just styling sugar, and don’t effect text selection. The adding of
“drag state” classes has been moved so it occurs on all devices. Touch
interfaces can now reflect when the slider UI is in use in a way
consistent with how the UI works on cursor based computers. No adverse
effects observed.